TIQR token enhancements
When we unlock the TIQR token with the App and use the wrong PIN/biometric code. Respond
back how many retries we have left before the token is locked/blocked

the TIQR app can not handle challenges bigger or equal then  `2^31 = 2147483648. Check if the
value is bigger if yes adjust it.